Make redo_grid use (nearly) same step size.
When using 'redo_grid' for cnoidal waves, use the current dx as the new
xStep. Therefore, the new dx will be approximately the same as the old
dx (up to rounding to ensure and integer number of steps fit between 0
and xLen).

Manually specify xStep and xLen to reproduce existing results.