Express H in terms of biviscosity nu_bi and eps/mu.
Explicitly factor out factors of mu/eps from the biviscosity nu_bi in
the numerical damping term H to retain our x -> lambda x and mu ->
mu/lambda^2 symmetry.

Note that we use different values of nu_bi to match the old values of H.
Likewise, we display the old values of nu_bi in plot legends and labels.
This ensures that all of the plots are unaltered.